+++ src/ua.rs
@@ -25,6 +25,7 @@ mod session_state;
 mod session_statistics;
 mod specified_attributes;
 mod subscription_id;
+mod unit_id;
 mod user_identity_token;
 
 #[cfg(feature = "mbedtls")]
@@ -50,6 +51,7 @@ pub use self::{
     session_statistics::SessionStatistics,
     specified_attributes::SpecifiedAttributes,
     subscription_id::SubscriptionId,
+    unit_id::UnitId,
     user_identity_token::UserIdentityToken,
 };
 pub(crate) use self::{client_config::ClientConfig, logger::Logger, server_config::ServerConfig};