+++ CHANGELOG.md
@@ -11,12 +11,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add method `Server::statistics()` to get server statistics.
 - Add `ua::ServerStatistics`, `ua::SecureChannelStatistics`, and `ua::SessionStatistics`.
+- Add `ua::EUInformation` and accompanying _newtype_ `ua::UnitId`.
+- Add `ua::Range`.
 
 ## [0.8.2] - 2025-03-19
 
 ### Added
 
-- Add wrapper data type `ua::Enumeration`.
+- Add `ua::Enumeration`.
 - Add `ValueType` variants `Structure` (inner type `ua::ExtensionObject`) and `Enumeration`.
 
 ### Changed